fix: update language switching
